Extract remote GitHub PR command builder
