+++ docs/DECISIONS.md
@@ -227,3 +227,4 @@
 - Git hunk patch execution lives in `GitPatchToolExecutor.swift`, and git tool errors live in `GitToolError.swift`. Hunk staging/restoring, patch temp-file handling, and patch-path metadata parsing should stay beside the patch executor; local and SSH Remote hunk planning should reuse `GitPatchToolExecutor.mismatchedPatchPath` instead of adding patch parsers to facades or remote planners.
 - Local git execution lives in `GitLocalToolExecutor.swift`. Status, diff, stage, restore, commit, push, path validation, and current-branch lookup should stay beside the local executor; `GitToolExecutor` remains only the compatibility facade over local, patch, worktree, and GitHub PR executors.
 - GitHub PR input validation and output parsing live outside PR command execution. `GitHubPullRequestInputValidator.swift` owns selector, reviewer, label, review-action, and merge-method normalization for both local and SSH Remote PR planning, while `GitHubPullRequestOutputParser.swift` owns URL artifact extraction. `GitHubPullRequestToolExecutor.swift` should stay focused on `gh pr` command construction and execution.
+- SSH Remote GitHub PR command construction lives in `WorkspaceRemoteGitHubPullRequestCommandBuilder.swift`. `WorkspaceRemoteGitToolRequestPlanner.swift` should route to that builder and keep generic git, hunk, worktree, and artifact intent planning separate so local PR validation, SSH Remote command assembly, and future QuillCloud remote transports do not drift.
